Refactor and implement Instances for external name services
