+++ test/integration/http-service.yaml.tmpl
@@ -1,3 +1,4 @@
+# Example service with an injected proxy
 apiVersion: v1
 kind: Service
 metadata:
